+++ sPHPf/vendors/coldstarstudios/framework/Error.php
@@ -4,6 +4,9 @@ namespace coldstarstudios\framework;
  * This class is used to create a debug inform system in the app framework.
  * The debug window will appear as a fixed menu in the bottom of the page
  * when an error is throwed.
+ * 
+ * All errors will add the 404 header because you could be looking for a page
+ * and probably dont exists. So the crawlers must avoid index this pages.
  *
  * @author Marcos Sigueros Fernández
  * @license MIT
@@ -28,6 +31,9 @@ class Error {
         $this->Application->data['error']['code'] = $this->exception->getCode();
         $this->Application->data['error']['message'] = $this->exception->getMessage();
         $this->Application->data['error']['file'] = $this->exception->getFile();
+        $this->Application->data['SERVER'] = $_SERVER;
+        
+        header('HTTP/1.0 404 Not Found');
         
         try{
             $no_production_view = new Response($this->Application->folders['error'], $this->Application->data);